+++ CHANGES.rst
@@ -217,6 +217,8 @@ Improvements
 - Allow filtering the list of editables by tags (:issue:`6195`, :pr:`6197`)
 - Warn users with a dialog before their session expires and let them extend it (:pr:`6026`,
   thanks :user:`SegiNyn`)
+- Add the option to include the picture of a registration in the conference participant list
+  (:pr:`6228`, thanks :user:`vtran99`)
 
 Bugfixes
 ^^^^^^^^